requirements.txt
# Core Framework & Server
fastapi==0.115.4
uvicorn[standard]==0.32.0

# GUI & Multimedia
PyQt6>=6.4.0
opencv-python>=4.8.0
psutil>=5.9.0

# AI & Cloud Services
azure-ai-inference>=1.0.0b1
azure-core>=1.30.0

# Voice Capabilities
SpeechRecognition>=3.14.1
pyttsx3==2.90
pyaudio==0.2.14
standard-aifc
audioop-lts

# Utilities & Config
requests==2.32.3
pydantic==2.9.2
python-dotenv==1.0.1
AppOpener==1.7

# Build Tools (Optional, for creating .exe)
pyinstaller>=6.0.0